fix(acp): isolate process protocol generations
